VoiceToText/requirements.txt
pvporcupine
pvrecorder
openai-whisper
numpy